requirements.txt
# Core dependencies
pinecone>=5.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Sparse embeddings
pinecone-text>=0.9.0

# Progress bars
tqdm>=4.65.0
